rename package & funcs
