+++ _internal/gapi-cloud-osconfig-go/cloud.google.com/go/osconfig/apiv1alpha1/os_config_client.go
@@ -21,7 +21,7 @@ import (
 	"math"
 	"time"
 
-	osconfigpb "github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/google-osconfig-agent/_internal/gapi-cloud-osconfig-go/google.golang.org/genproto/googleapis/cloud/osconfig/v1alpha1"
+	osconfigpb "github.com/GoogleCloudPlatform/osconfig/_internal/gapi-cloud-osconfig-go/google.golang.org/genproto/googleapis/cloud/osconfig/v1alpha1"
 	"github.com/golang/protobuf/proto"
 	gax "github.com/googleapis/gax-go"
 	"google.golang.org/api/iterator"
