+++ lock_linux.go
@@ -20,7 +20,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/google-osconfig-agent/logger"
+	"github.com/GoogleCloudPlatform/osconfig/logger"
 )
 
 func obtainLock() {
