+++ ospatch/system_windows.go
@@ -17,7 +17,7 @@
 package ospatch
 
 import (
-	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/google-osconfig-agent/logger"
+	"github.com/GoogleCloudPlatform/osconfig/logger"
 	"golang.org/x/sys/windows/registry"
 )
